Refactoring operation name : extract class
Refactoring object : MakeAlphaNumericToSegment

Reason : makeAlphanumeric, changeAlphaNumericStringToSegment has
independent role.
So, make these function to class(MakeAlphaNumericToSegment).
